# openapi.yaml
openapi: 3.0.0
info:
  title: Main Collection
  version: 1.0.0
  description: Contains the Postman docs for all of our api/endpoionts.
servers:
  - url: http://localhost:3000
paths:
...
  /api/threads/{threadId}/sentiment:
    post:
      summary: Create Sentiment
      responses: {}
      tags:
        - Threads
      security:
        - BearerAuth: []
      description: |-
        ## Create Sentiment Analysis

        **POST** `/api/threads/{id}/sentiment`

        Performs sentiment analysis on all posts in a match-associated thread.

        ### Headers

        ```
        Authorization: Bearer <access_token>

         ```

        ### Path Parameters

        - `id` (required): The ID of the thread to analyze
            

        ### Success Response

        **Code:** `200 OK`

        ``` json
        {
            "id": 5,
            "matchId": 12,
            "overall": 0.65,
            "homeTeam": 0.72,
            "awayTeam": -0.45,
            "numPosts": 150
        }

         ```

        ### Response Fields

        - `id`: Sentiment record ID
            
        - `matchId`: Associated match ID
            
        - `overall`: Overall sentiment score (-1 to 1)
            
        - `homeTeam`: Sentiment towards home team (-1 to 1)
            
        - `awayTeam`: Sentiment towards away team (-1 to 1)
            
        - `numPosts`: Number of posts analyzed
            

        ### Error Responses

        **Code:** `401 Unauthorized`

        ``` json
        {
            "error": "Unauthorized"
        }

         ```

        **Code:** `403 Forbidden`

        ``` json
        {
            "error": "User is banned"
        }

         ```

        **Code:** `400 Bad Request`

        ``` json
        {
            "error": "Invalid thread ID"
        }

         ```

        ``` json
        {
            "error": "Thread is not associated with a match"
        }

         ```

        ``` json
        {
            "error": "No posts found in thread"
        }

         ```

        **Code:** `404 Not Found`

        ``` json
        {
            "error": "Thread not found"
        }

         ```

        ``` json
        {
            "error": "Match not found"
        }

         ```

        ``` json
        {
            "error": "Teams not found"
        }

         ```

        **Code:** `500 Internal Server Error`

        ``` json
        {
            "error": "Failed to perform sentiment analysis"
        }

         ```

        ### Notes

        - AI model used: `cardiffnlp/twitter-roberta-base-sentiment-latest`
            
        - Sentiment scores: positive (1), negative (-1), neutral (0)
            
        - Posts are processed in batches to avoid input size limits
            
        - Maximum input length: 512 characters per batch
            
        - Batch size: 10 posts
    parameters:
      - name: threadId
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get Sentiment
      responses: {}
      tags:
        - Threads
      security:
        - BearerAuth: []
      description: |-
        ## Get Sentiment Analysis

        **GET** `/api/threads/{id}/sentiment`

        Retrieves existing sentiment analysis for a thread.

        ### Path Parameters

        - `id` (required): The ID of the thread
            

        ### Success Response

        **Code:** `200 OK`

        ``` json
        {
            "id": 5,
            "matchId": 12,
            "overall": 0.65,
            "homeTeam": 0.72,
            "awayTeam": -0.45,
            "numPosts": 150,
            "createdAt": "2026-03-06T10:00:00.000Z",
            "updatedAt": "2026-03-06T10:00:00.000Z"
        }

         ```

        ### Error Responses

        **Code:** `400 Bad Request`

        ``` json
        {
            "error": "Invalid thread ID"
        }

         ```

        ``` json
        {
            "error": "Thread is not associated with a match"
        }

         ```

        **Code:** `404 Not Found`

        ``` json
        {
            "error": "Thread not found"
        }

         ```

        ``` json
        {
            "error": "Sentiment analysis not found for this thread"
        }

         ```

        **Code:** `500 Internal Server Error`

        ``` json
        {
            "error": "Failed to fetch sentiment analysis"
        }

         ```
    put:
      summary: Update Sentiment
      responses: {}
      tags:
        - Threads
      security:
        - BearerAuth: []
      description: >-
        ## Update Sentiment Analysis


        **PUT** `/api/threads/{id}/sentiment`


        Updates sentiment analysis with new posts from the thread. Rate limited
        to once per hour.


        ### Path Parameters


        - `id` (required): The ID of the thread
            

        ### Success Response


        **Code:** `200 OK`


        ``` json

        {
            "id": 5,
            "matchId": 12,
            "overall": 0.68,
            "homeTeam": 0.75,
            "awayTeam": -0.42,
            "numPosts": 180,
            "updatedAt": "2026-03-06T11:30:00.000Z"
        }

         ```

        ### Error Responses


        **Code:** `400 Bad Request`


        ``` json

        {
            "error": "Invalid thread ID"
        }

         ```

        ``` json

        {
            "error": "Thread is not associated with a match"
        }

         ```

        **Code:** `404 Not Found`


        ``` json

        {
            "error": "Thread not found"
        }

         ```

        ``` json

        {
            "error": "Match not found"
        }

         ```

        ``` json

        {
            "error": "Teams not found"
        }

         ```

        **Code:** `429 Too Many Requests`


        ``` json

        {
            "error": "Sentiment analysis was updated less than an hour ago. Please try again later."
        }

         ```

        **Code:** `500 Internal Server Error`


        ``` json

        {
            "error": "Failed to update sentiment analysis"
        }

         ```

        ### Notes


        - Only analyzes posts created since the last update
            
        - If no new posts exist, returns the existing sentiment data
            
        - Uses weighted average based on post counts
            
        - Rate limited to prevent abuse: 1 update per hour
      requestBody:
        content:
          application/json:
            schema:
              type: object
    delete:
      summary: Delete Sentiment
      responses: {}
      tags:
        - Threads
      security:
        - BearerAuth: []
      description: |-
        ## Delete Sentiment Analysis

        **DELETE** `/api/threads/{id}/sentiment`

        Deletes sentiment analysis for a thread.

        ### Path Parameters

        - `id` (required): The ID of the thread
            

        ### Success Response

        **Code:** `200 OK`

        ``` json
        {
            "message": "Sentiment analysis deleted successfully"
        }

         ```

        ### Error Responses

        **Code:** `400 Bad Request`

        ``` json
        {
            "error": "Invalid thread ID"
        }

         ```

        ``` json
        {
            "error": "Thread is not associated with a match"
        }

         ```

        **Code:** `404 Not Found`

        ``` json
        {
            "error": "Thread not found"
        }

         ```

        **Code:** `500 Internal Server Error`

        ``` json
        {
            "error": "Failed to delete sentiment analysis"
        }

         ```
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
security:
  - BearerAuth: []